requirements.txt
google-generativeai>=0.3.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
PyQt5>=5.15.0
pyaudio>=0.2.13
pydub>=0.25.1
psutil>=5.9.0
pygetwindow>=0.0.9
pillow>=10.0.0
numpy>=1.24.0
